ProcessProvider implementation for Windows
This adds support for obtaining information about arbitrary processes
under Windows OS. Previously, it was possible to fetch information about
the current process only.

This uses some undocumented APIs to fetch command-line arguments and
current working directory.

It is meant to work in XP onwards. As a 32 bit process under a 32 bit OS
or a 64 bit process in a 64 bit OS. Some features will be missing when
running as a 32-bit process in a 64-bit OS (WOW64): Command-line
arguments and cwd.